Integrate takeoff form
